+++ src/Ch03/LocalFunction.hs
@@ -1,0 +1,20 @@
+-- |
+-- Module      :  Ch03.LocalFunction
+-- Description :  Local Functions, Global Variables
+-- Copyright   :  erlnow 2020 - 2030
+-- License     :  BSD3
+--
+-- Maintainer  :  erlestau@gmail.com
+-- Stability   :  experimental
+-- Portability :  unknown
+--
+-- Example from Chapter 3: Defining Types, Streamlining Functions
+
+module Ch03.LocalFunction where
+
+-- Local Functions, Global Variables, p. 64
+pluralise :: String -> [Int] -> [String]
+pluralise word counts = map plural counts
+  where plural 0 = "no " ++ word ++ "s"
+        plural 1 = "one " ++ word
+        plural n = show n ++ " " ++ word ++ "s"